Rework HomeWork Part One
