+++ src/by/ruslan/radevich/homework/lessons3/FirstTask.java
@@ -1,8 +1,8 @@
-package by.ruslan.radevich.homework.lessonsThird;
+package by.ruslan.radevich.homework.lessons3;
 
 import java.util.Scanner;
 
-public class TaskOne {
+public class FirstTask {
     public static void main(String[] args) {
         Scanner input = new Scanner(System.in);
         System.out.println("Enter the number color");
