+++ src/by/ruslan/radevich/homework/lessons3/FourthTask.java
@@ -1,15 +1,17 @@
-package by.ruslan.radevich.homework.lessonsThird;
+package by.ruslan.radevich.homework.lessons3;
 
 import java.util.Random;
 import java.util.Scanner;
 
-public class TaskFour {
+public class FourthTask {
     public static void main(String[] args) {
         Scanner input = new Scanner(System.in);
         System.out.println("Enter your number");
         int number = input.nextInt();
+        while (number < 3) {
+            number += input.nextInt();
+        }
         Random numbers = new Random();
-        for (int i = number; i < 5; i++) ;
         System.out.println("Random numbers " + numbers.nextInt(5));
